Bug fixed in operators

+++ src/operators/operators_mutation.hpp
@@ -3,7 +3,7 @@
 #include <random>
 
 using T = std::vector<std::vector<int>>;
-using L = int;
+using L = double;
 
 // Mutation Operators ---------------------------------------------------------------
